+++ app/ante/cosmos/fees_test.go
@@ -63,14 +63,14 @@ func (suite *AnteTestSuite) TestDeductFeeDecorator() {
 			errContains: "must provide positive gas",
 		},
 		{
-			name:        "fail - checkTx - insufficient funds and no staking rewards",
+			name:        "fail - checkTx - insufficient funds",
 			balance:     zero,
 			rewards:     []math.Int{zero},
 			gas:         10_000_000,
 			checkTx:     true,
 			simulate:    false,
 			expPass:     false,
-			errContains: "insufficient funds and failed to claim sufficient staking rewards",
+			errContains: "failed to deduct fee: spendable balance",
 			postCheck: func() {
 				// the balance should not have changed
 				balance := suite.app.BankKeeper.GetBalance(suite.ctx, addr, utils.BaseDenom)
@@ -83,37 +83,14 @@ func (suite *AnteTestSuite) TestDeductFeeDecorator() {
 			},
 		},
 		{
-			name:        "pass - insufficient funds but sufficient staking rewards",
-			balance:     zero,
-			rewards:     []math.Int{initBalance},
-			gas:         10_000_000,
-			checkTx:     false,
-			simulate:    false,
-			expPass:     true,
-			errContains: "",
-			postCheck: func() {
-				// the balance should have increased
-				balance := suite.app.BankKeeper.GetBalance(suite.ctx, addr, utils.BaseDenom)
-				suite.Require().False(
-					balance.Amount.IsZero(),
-					"expected balance to have increased after withdrawing a surplus amount of staking rewards",
-				)
-
-				// the rewards should all have been withdrawn
-				rewards, err := testutil.GetTotalDelegationRewards(suite.ctx, suite.app.DistrKeeper, addr)
-				suite.Require().NoError(err, "failed to get total delegation rewards")
-				suite.Require().Empty(rewards, "expected all rewards to be withdrawn")
-			},
-		},
-		{
-			name:        "fail - insufficient funds and insufficient staking rewards",
+			name:        "fail - insufficient funds",
 			balance:     math.NewInt(1e5),
 			rewards:     []math.Int{math.NewInt(1e5)},
 			gas:         10_000_000,
 			checkTx:     false,
 			simulate:    false,
 			expPass:     false,
-			errContains: "insufficient funds and failed to claim sufficient staking rewards",
+			errContains: "failed to deduct fee: spendable balance",
 			postCheck: func() {
 				// the balance should not have changed
 				balance := suite.app.BankKeeper.GetBalance(suite.ctx, addr, utils.BaseDenom)